Refactor: 스타일 파일이름과 페이지 이름을 일치
